backend/requirements/base.txt
Django>=5.0,<5.1
djangorestframework>=3.15,<3.16
django-cors-headers>=4.3,<4.4
django-filter>=24.0,<25.0
drf-spectacular>=0.27,<0.28
psycopg2-binary>=2.9,<2.10
python-decouple>=3.8,<3.9
gunicorn>=22.0,<23.0
Pillow>=10.2,<11.0
djangorestframework-simplejwt>=5.3,<5.4
argon2-cffi>=23.1,<24.0
